feat: better pagination types
