+++ packages/pagination/PaginationItem.stories.tsx
@@ -1,8 +1,5 @@
-import React from 'react'
 import { Story, Meta } from '@storybook/react'
-
-import { PaginationItemVariant, PaginationItemProps } from './types'
-import PaginationItem from './PaginationItem'
+import { PaginationItem, PaginationItemVariant, PaginationItemProps } from '.'
 
 export default {
   component: PaginationItem,
