+++ packages/pagination/PaginationItem.tsx
@@ -1,11 +1,19 @@
-import React from 'react'
-import { PaginationItemStyle } from './PaginationItemStyles'
-import { PaginationItemProps, PaginationItemVariant } from './types'
+import { FC, ReactNode } from 'react'
+import { PaginationItemStyle, PaginationItemStyleProps } from './PaginationItemStyles'
+import { PaginationItemVariant, PaginationItemVariants } from './constants'
 
-const PaginationItem: React.FC<PaginationItemProps> = (props) => {
-  const { icon, variant = PaginationItemVariant.default, ...rest } = props
+export type PaginationItemProps = Omit<PaginationItemStyleProps, '$variant'> & {
+  variant?: PaginationItemVariants
+  icon: ReactNode
+}
+
+const PaginationItem: FC<PaginationItemProps> = ({
+  icon,
+  variant = PaginationItemVariant.default,
+  ...rest
+}) => {
   return (
-    <PaginationItemStyle $variant={variant} {...rest}>
+    <PaginationItemStyle {...rest} $variant={variant}>
       {icon}
     </PaginationItemStyle>
   )
