+++ packages/pagination/constants.ts
@@ -1,0 +1,5 @@
+export enum PaginationItemVariant {
+  default = 'default',
+  active = 'active',
+}
+export type PaginationItemVariants = keyof typeof PaginationItemVariant
